"First" commit
This is the second version of ledge, rewritten to be async without the
use of celery (via twisted).

I did quite a bit of work locally on this iteration before pushing, and
then rehoused it into my newest python project template.

Hopefully too many things didnt get jumbled around in the move(s).