Minor Refactors
Removes the callInThread_or_callLater function, as it relies on an
interface which will only exist on handler implementations.

The associated logic now lives in the `.process` method handlers.

Made the name of the variable the dictates thread safety a positive
phrasing, so it would be clearer. Added a default truth-y value to the
handler implementation base.